fix(time): correct native timeval microseconds ABI
Linux LP64 timeval contains two signed 64-bit fields. PosixSusecondsT was c_int, leaving four bytes of trailing padding even though the structure still had the expected 16-byte size. Copying that structure to userspace exposed uninitialized padding and made V8 interpret the microseconds field as a large negative number, moving Node Date values back by years.

Use c_long for native microseconds and assert field width, offsets, and timeval/itimerval sizes at compile time. Update producers to write the complete field. Validate full-width setitimer and utimes microseconds before narrowing or conversion, preserving the distinction between timer intervals and signed file timestamps.

Handle widened select input with checked normalization and wide intermediate arithmetic, retaining the existing millisecond granularity and bounding deadlines at the existing Instant representation limit. This does not introduce a new polling framework or claim full Linux equivalence for extreme timeout writeback.

Add and register eight focused dunitest cases covering output completeness, realtime agreement, high-bit input validation, select deadline bounds, and utimes metadata preservation.

Validation: seven new cases failed on the original kernel (the extreme select case was not run there); all eight passed on Linux and on the fixed DragonOS kernel. A second guest boot passed ten suite repetitions (80 executions). The focused existing suites passed 42 cases before and after, with two optional stress tests skipped. Unmodified Node passed ten Date.now/date bracket checks across two boots. GDB and disassembly confirmed full-width initialized microseconds output. make kernel, targeted rustfmt checks, and diff checks passed.

+++ kernel/src/filesystem/poll.rs
@@ -418,7 +418,7 @@ pub fn poll_select_finish(
         PollTimeType::TimeVal | PollTimeType::OldTimeVal => {
             let rtv = PosixTimeval {
                 tv_sec: rts.tv_sec,
-                tv_usec: (rts.tv_nsec / 1000) as i32,
+                tv_usec: rts.tv_nsec / 1000,
             };
             let mut tvwriter = UserBufferWriter::new(
                 user_time_ptr as *mut PosixTimeval,
